+++ pylf/tests/test_page.py
@@ -1,7 +1,7 @@
 # coding: utf-8
 import PIL.Image
 
-from pylf import _page
+from pylf._page import *
 
 
 def test_image():
@@ -9,13 +9,13 @@ def test_image():
     size = (1, 1)
     color = "white"
     im = PIL.Image.new(mode, size, color)
-    page = _page.Page(mode, size, color, 0)
+    page = Page(mode, size, color, 0)
     assert page.image == im
 
 
 def test_num():
     num = 3
-    page = _page.Page("L", (1, 1), "white", num)
+    page = Page("L", (1, 1), "white", num)
     assert page.num == num
 
 
@@ -31,17 +31,17 @@ def test_matrix():
 
 def test_size():
     size = (1, 2)
-    page = _page.Page("CMYK", size, "white", 0)
+    page = Page("CMYK", size, "white", 0)
     assert page.size() == size
 
 
 def test_width():
     size = (1, 2)
-    page = _page.Page("CMYK", size, "white", 0)
+    page = Page("CMYK", size, "white", 0)
     assert page.width() == size[0]
 
 
 def test_height():
     size = (1, 2)
-    page = _page.Page("CMYK", size, "white", 0)
+    page = Page("CMYK", size, "white", 0)
     assert page.height() == size[1]