+++ tests/watch.py
@@ -10,39 +10,48 @@ from tests.util import *
 SEED = "PyLf"
 
 
-def test_handwrite2():
+def main():
+    print(
+        "Test by naked eyes:\n"
+        "======================================\n"
+        "Test: pylf.handwrite2"
+    )
     path = "backgrounds/even-odd-letter/"
     image1 = PIL.Image.open(abs_path(path + "村庄信笺纸.jpg"))
     image2 = PIL.Image.open(abs_path(path + "树信笺纸.jpg"))
     assert image1.mode == "RGB"
     assert image2.mode == "RGB"
-    images = (image1, image2)
 
-    template2 = {
-        "backgrounds": [im.resize(size=(im.size[0] * 2, im.size[1] * 2)) for im in images],
-        "margins": (
-            {"left": 40, "top": 200, "right": 30, "bottom": 560},
-            {"left": 40, "top": 200, "right": 30, "bottom": 980},
-        ),
-        "line_spacings": (88, 88),
-        "font_sizes": (74, 74),
-        "font": get_default_font(),
-        "fill": (0, 0, 0),
-    }
+    template1 = Template(
+        background=image1.resize(size=(image1.size[0] * 2, image1.size[1] * 2)),
+        left_margin=40,
+        top_margin=200,
+        right_margin=30,
+        bottom_margin=560,
+        line_spacing=88,
+        font_size=74,
+        font=get_default_font(),
+    )
+    template2 = Template(
+        background=image2.resize(size=(image2.size[0] * 2, image2.size[1] * 2)),
+        left_margin=40,
+        top_margin=200,
+        right_margin=30,
+        bottom_margin=980,
+        line_spacing=88,
+        font_size=74,
+        font=get_default_font(),
+    )
+    templates = (template1, template2)
     for file in pathlib.Path(abs_path("texts")).iterdir():
         print(file)
         with file.open(encoding="utf-8") as f:
             text = f.read()
-        images = handwrite2(text, template2, seed=SEED)
+        images = handwrite(text, templates, seed=SEED)
         for im in images:
             im.show()
         assert input("Like it? [Y/N] ").upper() == "Y"
 
 
 if __name__ == "__main__":
-    print(
-        "Test by naked eyes:\n"
-        "======================================\n"
-        "Test: pylf.handwrite2"
-    )
-    test_handwrite2()
+    main()