+++ docs/release_notes.md
@@ -8,7 +8,12 @@
 * __将参数`color`替换为新参数`fill`__
 * __取消背景图片的`mode`限制，提供对全部`mode`的支持。__
 * 取消`worker`和`seed`必须以`keyword`形式传入的限制
-* __创建PyLf的基异常类`Error`，并将`LayoutError`改为直接继承自该异常类。__
+* 创建PyLf的基异常类`Error`，并将`LayoutError`改为直接继承自该异常类。
+* 创建新异常类`BackgroundTooLargeError`
+* __添加public module `pylf.core`__
+* __创建模板类`Template`，用于管理模板相关参数。__
+* __将`handwrite2()`合并入`handwrite()`__
+* __修改`handwrite()`的接口，使其接收`Template`类的实例。__
 
 ## v3.4.0 (2019-5-1)
 * __Pillow版本限制放宽至`>= 5, < 7`__